Settings executions list, generic dialog component and small improvements to settings pages
