+++ frontend/src/assets/scss/dialogs.scss
@@ -32,6 +32,6 @@
   }
 
   &__footer {
-    @apply bg-gray-50 flex items-center px-6 py-4 h-fit rounded-b-lg justify-between;
+    @apply bg-gray-50 flex items-center px-6 py-4 h-fit rounded-b-lg;
   }
 }
